+++ src/lib/series/legend.ts
@@ -57,7 +57,7 @@ export function computeLegend(
       isLegendItemVisible: !hideInLegend,
       displayValue: {
         raw: series.lastValue,
-        formatted: formatter(series.lastValue),
+        formatted: isSeriesVisible ? formatter(series.lastValue) : undefined,
       },
     });
   });