+++ src/lib/series/stacked_series_utils.ts
@@ -1,5 +1,11 @@
 import { DataSeries, DataSeriesDatum, RawDataSeries } from './series';
 
+interface StackedValues {
+  values: number[];
+  percent: number[];
+  total: number;
+}
+
 /**
  * Map each y value from a RawDataSeries on it's specific x value into,
  * ordering the stack based on the dataseries index.
@@ -8,7 +14,7 @@ import { DataSeries, DataSeriesDatum, RawDataSeries } from './series';
 export function getYValueStackMap(dataseries: RawDataSeries[]): Map<any, number[]> {
   const stackMap = new Map<any, number[]>();
   dataseries.forEach((ds, index) => {
-    for (const [group, data] of ds.data) {
+    for (const [, data] of ds.data) {
       data.forEach((datum) => {
         const stack = stackMap.get(datum.x) || new Array(dataseries.length).fill(0);
         stack[index] = datum.y1;
@@ -28,28 +34,51 @@ export function getYValueStackMap(dataseries: RawDataSeries[]): Map<any, number[
 export function computeYStackedMapValues(
   yValueStackMap: Map<any, number[]>,
   scaleToExtent: boolean,
-): Map<any, number[]> {
-  const stackedValues = new Map<any, number[]>();
+): Map<any, StackedValues> {
+  const stackedValues = new Map<any, StackedValues>();
 
   yValueStackMap.forEach((yStackArray, xValue) => {
     const stackArray = yStackArray.reduce(
-      (stackedValue, currentValue, index) => {
-        if (stackedValue.length === 0) {
+      (acc, currentValue, index) => {
+        if (acc.values.length === 0) {
           if (scaleToExtent) {
-            return [currentValue, currentValue];
+            return {
+              values: [currentValue, currentValue],
+              total: currentValue,
+            };
           }
-          return [0, currentValue];
+          return {
+            values: [0, currentValue],
+            total: currentValue,
+          };
         }
-        return [...stackedValue, stackedValue[index] + currentValue];
+        return {
+          values: [...acc.values, acc.values[index] + currentValue],
+          total: acc.total + currentValue,
+        };
       },
-      [] as number[],
+      {
+        values: [] as number[],
+        total: 0,
+      },
     );
-    stackedValues.set(xValue, stackArray);
+    const percent = stackArray.values.map((value) => {
+      return value / stackArray.total;
+    });
+    stackedValues.set(xValue, {
+      values: stackArray.values,
+      percent,
+      total: stackArray.total,
+    });
   });
   return stackedValues;
 }
 
-export function formatStackedDataSeriesValues(dataseries: RawDataSeries[], scaleToExtent: boolean): DataSeries[] {
+export function formatStackedDataSeriesValues(
+  dataseries: RawDataSeries[],
+  scaleToExtent: boolean,
+  isPercentageMode: boolean = false,
+): DataSeries[] {
   const yValueStackMap = getYValueStackMap(dataseries);
 
   const stackedValues = computeYStackedMapValues(yValueStackMap, scaleToExtent);
@@ -58,17 +87,18 @@ export function formatStackedDataSeriesValues(dataseries: RawDataSeries[], scale
     const newData = new Map<string, DataSeriesDatum[]>();
     for (const [group, series] of ds.data) {
       series.forEach((data) => {
-        const { x, y1, datum, xAccessor, yAccessor } = data;
+        const { x, y1, y0, datum, xAccessor, yAccessor } = data;
         if (stackedValues.get(x) === undefined) {
           return;
         }
+
         let computedY0: number | null;
         if (scaleToExtent) {
-          computedY0 = data.y0 ? data.y0 : y1;
+          computedY0 = y0 ? y0 : y1;
         } else {
-          computedY0 = data.y0 ? data.y0 : 0;
+          computedY0 = y0 ? y0 : 0;
         }
-        const initialY0 = data.y0 == null ? null : data.y0;
+        const initialY0 = y0 == null ? null : y0;
         if (seriesIndex === 0) {
           if (!newData.has(group)) {
             newData.set(group, []);
@@ -89,13 +119,20 @@ export function formatStackedDataSeriesValues(dataseries: RawDataSeries[], scale
           if (!stack) {
             return;
           }
-          const stackY = stack[seriesIndex];
-          const stackedY1 = y1 !== null ? stackY + y1 : null;
-          let stackedY0: number | null = data.y0 == null ? stackY : stackY + data.y0;
-          // configure null y0 if y1 is null
-          // it's semantically right to say y0 is null if y1 is null
-          if (stackedY1 === null) {
-            stackedY0 = null;
+          const stackY = isPercentageMode ? stack.percent[seriesIndex] : stack.values[seriesIndex];
+          let stackedY1: number | null = null;
+          let stackedY0: number | null = null;
+          if (isPercentageMode) {
+            stackedY1 = y1 !== null ? stackY + y1 : null;
+            stackedY0 = y0 != null ? stackY + y0 : stackY;
+          } else {
+            stackedY1 = y1 !== null ? stackY + y1 : null;
+            stackedY0 = y0 != null ? stackY + y0 : stackY;
+            // configure null y0 if y1 is null
+            // it's semantically right to say y0 is null if y1 is null
+            if (stackedY1 === null) {
+              stackedY0 = null;
+            }
           }
 
           if (!newData.has(group)) {